[FEAT] profile → readinglog로 명칭 변경 및 경로 수정 #3
좀 더 기록한다는 느낌이 들도록 변경
