+++ BookHub/src/main/resources/static/css/readinglog/edit.css
@@ -1,1 +1,40 @@
 @charset "UTF-8";
+
+html,
+body {
+    margin: 0;
+    padding: 0;
+}
+
+#readinglog-wrapper {
+    box-sizing: border-box;
+    width: 940px;
+    margin: 0 auto;
+    margin-top: 80px;
+}
+
+#readinglog-head {
+    box-sizing: border-box;
+    width: 100%;
+    height: 200px;
+    padding: 20px;
+    background-color: black;
+}
+
+#bookimage {
+    float: left;
+    width: 160px;
+    height: 160px;
+    object-fit: contain;
+}
+#bookTitle {
+    color: white;
+}
+#bookAuthor {
+    color: white;
+}
+
+#readinglog-body {
+    box-sizing: border-box;
+    width: 100%;
+}
