+++ BookHub/src/main/webapp/WEB-INF/views/readinglog/edit.jsp
@@ -6,7 +6,6 @@
     <head>
         <meta charset="UTF-8">
         <title>BookHub :: 독서 활동 기록하기</title>
-        <link rel="stylesheet" href="/static/css/readinglog/readinglog.css">
         <link rel="stylesheet" href="/static/css/readinglog/edit.css">
     </head>
